+++ src/app/app.module.ts
@@ -29,6 +29,8 @@ import { PwaService } from './services/pwa.service';
 import { environment } from '../environments/environment';
 import { PortalModule } from './pages';
 import { AppComponent } from './app.component';
+import { HeaderModule } from './pages/header/header.module';
+import { FooterModule } from './pages/footer/footer.module';
 import { ServiceWorkerModule } from '@angular/service-worker';
 
 @NgModule({
@@ -44,6 +46,8 @@ import { ServiceWorkerModule } from '@angular/service-worker';
     IgoStopPropagationModule,
     PortalModule,
     HammerModule,
+    HeaderModule,
+    FooterModule,
     ServiceWorkerModule.register('ngsw-worker.js', { enabled: environment.production })
   ],
   providers: [