+++ src/app/pages/portal/portal.component.ts
@@ -62,9 +62,16 @@ import {
   QueryService,
   WfsWorkspace,
   FeatureWorkspace,
+  EditionWorkspace,
+  EditionWorkspaceService,
   generateIdFromSourceOptions,
   computeOlFeaturesExtent,
   addStopToStore,
+  ImageLayer,
+  VectorLayer,
+  MapExtent,
+  moveToOlFeatures,
+  FeatureMotion,
   XYZDataSource,
   FeatureStore,
   VectorLayer,
@@ -129,13 +136,15 @@ export class PortalComponent implements OnInit, OnDestroy {
   public minSearchTermLength = 2;
   public hasExpansionPanel = false;
   public hasGeolocateButton = true;
+  public hasHomeExtentButton = false;
+  public showMenuButton = true;
+  public showSearchBar = true;
   public showRotationButtonIfNoRotation = false;
   public hasFeatureEmphasisOnSelection: Boolean = false;
   public workspaceNotAvailableMessage: String = 'workspace.disabled.resolution';
   public workspacePaginator: MatPaginator;
-  public workspaceEntitySortChange$: BehaviorSubject<
-    boolean
-  > = new BehaviorSubject(false);
+  public workspaceEntitySortChange$: BehaviorSubject<boolean> = new BehaviorSubject(false);
+  public workspaceSwitchDisabled = false;
   public paginatorOptions: EntityTablePaginatorOptions = {
     pageSize: 50, // Number of items to display on a page.
     pageSizeOptions: [1, 5, 10, 20, 50, 100, 500] // The set of provided page size options to display to the user.
@@ -179,6 +188,9 @@ export class PortalComponent implements OnInit, OnDestroy {
 
   public offlineRegionStore: FeatureStore = new FeatureStore([], {map: this.map});
 
+  public homeExtent: MapExtent;
+  public homeCenter: [number, number];
+  public homeZoom: number;
   @ViewChild('mapBrowser', { read: ElementRef, static: true })
   mapBrowser: ElementRef;
   @ViewChild('searchBar', { read: ElementRef, static: true })
@@ -241,6 +253,11 @@ export class PortalComponent implements OnInit, OnDestroy {
   }
   set expansionPanelExpanded(value: boolean) {
     this.workspaceState.workspacePanelExpanded = value;
+    if (value === true) {
+      this.map.viewController.setPadding({bottom: 280});
+    } else {
+      this.map.viewController.setPadding({bottom: 0});
+    }
   }
 
   get toastPanelShown(): boolean {
@@ -304,6 +321,7 @@ export class PortalComponent implements OnInit, OnDestroy {
     return this.workspaceState.workspace$.value;
   }
 
+
   constructor(
     private route: ActivatedRoute,
     public workspaceState: WorkspaceState,
@@ -328,25 +346,26 @@ export class PortalComponent implements OnInit, OnDestroy {
     public dialogWindow: MatDialog,
     private queryService: QueryService,
     private storageService: StorageService,
+    private editionWorkspaceService: EditionWorkspaceService,
     private downloadState: DownloadState,
     private directionState: DirectionState
   ) {
     this.hasExpansionPanel = this.configService.getConfig('hasExpansionPanel');
-    this.hasGeolocateButton =
-      this.configService.getConfig('hasGeolocateButton') === undefined ? true : this.configService.getConfig('hasGeolocateButton');
-    this.showRotationButtonIfNoRotation =
-      this.configService.getConfig('showRotationButtonIfNoRotation') === undefined ?
-        false :
-        this.configService.getConfig('showRotationButtonIfNoRotation');
+    this.hasHomeExtentButton =
+      this.configService.getConfig('homeExtentButton') === undefined ? false : true;
+    this.hasGeolocateButton = this.configService.getConfig('hasGeolocateButton') === undefined ? true :
+      this.configService.getConfig('hasGeolocateButton');
+    this.showRotationButtonIfNoRotation = this.configService.getConfig('showRotationButtonIfNoRotation') === undefined ? false :
+      this.configService.getConfig('showRotationButtonIfNoRotation');
+    this.showMenuButton = this.configService.getConfig('showMenuButton') === undefined ? true :
+      this.configService.getConfig('showMenuButton');
+    this.showSearchBar = this.configService.getConfig('showSearchBar') === undefined ? true :
+      this.configService.getConfig('showSearchBar');
     this.forceCoordsNA = this.configService.getConfig('app.forceCoordsNA');
-    this.hasFeatureEmphasisOnSelection = this.configService.getConfig(
-      'hasFeatureEmphasisOnSelection'
-    );
+    this.hasFeatureEmphasisOnSelection = this.configService.getConfig('hasFeatureEmphasisOnSelection');
 
 
-    this.igoSearchPointerSummaryEnabled = this.configService.getConfig(
-      'hasSearchPointerSummary'
-    );
+    this.igoSearchPointerSummaryEnabled = this.configService.getConfig('hasSearchPointerSummary');
     if (this.igoSearchPointerSummaryEnabled === undefined) {
       this.igoSearchPointerSummaryEnabled = this.storageService.get('searchPointerSummaryEnabled') as boolean || false;
     }
@@ -366,30 +385,31 @@ export class PortalComponent implements OnInit, OnDestroy {
       (context: DetailedContext) => this.onChangeContext(context)
     );
 
-    this.contextMenuStore.load([
-      {
-        id: 'download',
-        title: 'download',
-        handler: () => {
-          this.onBeforeDownload();
-        }
-      },
-      {
-        id: 'coordinates',
-        title: 'coordinates',
-        handler: () => this.searchCoordinate(this.contextMenuCoord)
-      },
-      {
-        id: 'googleMaps',
-        title: 'googleMap',
-        handler: () => this.openGoogleMaps(this.contextMenuCoord)
-      },
-      {
-        id: 'googleStreetView',
-        title: 'googleStreetView',
-        handler: () => this.openGoogleStreetView(this.contextMenuCoord)
+    const contextActions = [
+    {
+      id: 'download',
+      title: 'download',
+      handler: () => {
+        this.onBeforeDownload();
       }
-    ]);
+    },
+    {
+      id: 'coordinates',
+      title: 'coordinates',
+      handler: () => this.searchCoordinate(this.contextMenuCoord)
+    },
+    {
+      id: 'googleMaps',
+      title: 'googleMap',
+      handler: () => this.openGoogleMaps(this.contextMenuCoord)
+    },
+    {
+      id: 'googleStreetView',
+      title: 'googleStreetView',
+      handler: () => this.openGoogleStreetView(this.contextMenuCoord)
+    }];
+
+    this.contextMenuStore.load(contextActions);
 
     this.queryStore.count$
       .pipe(pairwise())
@@ -468,10 +488,22 @@ export class PortalComponent implements OnInit, OnDestroy {
       this.workspaceMaximize$.subscribe(() => this.updateMapBrowserClass())
     );
 
-    this.workspaceState.workspace$.subscribe((activeWks: WfsWorkspace | FeatureWorkspace) => {
+    this.workspaceState.workspace$.subscribe((activeWks: WfsWorkspace | FeatureWorkspace | EditionWorkspace) => {
       if (activeWks) {
         this.selectedWorkspace$.next(activeWks);
         this.expansionPanelExpanded = true;
+
+        if (activeWks.layer.options.workspace?.pageSize && activeWks.layer.options.workspace?.pageSizeOptions) {
+          this.paginatorOptions = {
+            pageSize: activeWks.layer.options.workspace?.pageSize,
+            pageSizeOptions: activeWks.layer.options.workspace?.pageSizeOptions
+          };
+        } else {
+          this.paginatorOptions = {
+            pageSize: 50,
+            pageSizeOptions: [1, 5, 10, 20, 50, 100, 500]
+          };
+        }
       } else {
         this.expansionPanelExpanded = false;
       }
@@ -524,7 +556,7 @@ export class PortalComponent implements OnInit, OnDestroy {
   }
 
   workspaceVisibility(): boolean {
-    const wks = (this.selectedWorkspace$.value as WfsWorkspace | FeatureWorkspace);
+    const wks = (this.selectedWorkspace$.value as WfsWorkspace | FeatureWorkspace | EditionWorkspace);
     if (wks.inResolutionRange$.value) {
       if (wks.entityStore.empty$.value && !wks.layer.visible) {
         this.workspaceNotAvailableMessage = 'workspace.disabled.visible';
@@ -537,6 +569,32 @@ export class PortalComponent implements OnInit, OnDestroy {
     return wks.inResolutionRange$.value;
   }
 
+  isEditionWorkspace(workspace) {
+    if (workspace instanceof EditionWorkspace) {
+      return true;
+    }
+    return false;
+  }
+
+  addFeature(workspace: EditionWorkspace) {
+    let feature = {
+      type: "Feature",
+      properties: {}
+    };
+    feature.properties = this.createFeatureProperties(workspace.layer);
+    this.workspaceState.rowsInMapExtentCheckCondition$.next(false);
+    workspace.editFeature(feature, workspace);
+  }
+
+  createFeatureProperties(layer: ImageLayer | VectorLayer) {
+    let properties = {};
+    layer.options.sourceOptions.sourceFields.forEach(field => {
+      if (!field.primary && field.visible) {
+        properties[field.name] = '';
+      }
+    });
+    return properties;
+  }
 
   paginatorChange(matPaginator: MatPaginator) {
     this.workspacePaginator = matPaginator;
@@ -549,6 +607,10 @@ export class PortalComponent implements OnInit, OnDestroy {
   entitySelectChange(result: { added: Feature[] }) {
     const baseQuerySearchSource = this.getQuerySearchSource();
     const querySearchSourceArray: QuerySearchSource[] = [];
+
+    if (this.selectedWorkspace$.value instanceof WfsWorkspace || this.selectedWorkspace$.value instanceof FeatureWorkspace) {
+      if (!this.selectedWorkspace$.value.getLayerWksOptionTabQuery()) {return;}
+    }
     if (result && result.added) {
       const results = result.added.map((res) => {
         if (
@@ -634,11 +696,15 @@ export class PortalComponent implements OnInit, OnDestroy {
   onMapQuery(event: { features: Feature[]; event: MapBrowserEvent<any> }) {
     const baseQuerySearchSource = this.getQuerySearchSource();
     const querySearchSourceArray: QuerySearchSource[] = [];
-
     const results = event.features.map((feature: Feature) => {
       let querySearchSource = querySearchSourceArray.find(
         (s) => s.title === feature.meta.sourceTitle
       );
+      if (this.getFeatureIsSameActiveWks(feature)) {
+        if (this.getWksActiveOpenInResolution() && !(this.workspace as WfsWorkspace).getLayerWksOptionMapQuery()) {
+          return;
+        }
+      }
       if (!querySearchSource) {
         querySearchSource = new QuerySearchSource({
           title: feature.meta.sourceTitle
@@ -647,9 +713,9 @@ export class PortalComponent implements OnInit, OnDestroy {
       }
       return featureToSearchResult(feature, querySearchSource);
     });
-
+    const filteredResults = results.filter(x => x !== undefined);
     const research = {
-      request: of(results),
+      request: of(filteredResults),
       reverse: false,
       source: baseQuerySearchSource
     };
@@ -722,9 +788,22 @@ export class PortalComponent implements OnInit, OnDestroy {
   }
 
   public toolChanged(tool: Tool) {
-    if (tool && tool.name === 'searchResults') {
+    if (tool && tool.name === 'searchResults' && this.searchBar) {
       this.searchBar.nativeElement.getElementsByTagName('input')[0].focus();
     }
+  }
+
+  private computeHomeExtentValues(context: DetailedContext) {
+    if (context?.map?.view?.homeExtent) {
+      this.homeExtent = context.map.view.homeExtent.extent;
+      this.homeCenter = context.map.view.homeExtent.center;
+      this.homeZoom = context.map.view.homeExtent.zoom;
+    } else {
+      this.homeExtent = undefined;
+      this.homeCenter = undefined;
+      this.homeZoom = undefined;
+    }
+
   }
 
   private onChangeContext(context: DetailedContext) {
@@ -732,6 +811,14 @@ export class PortalComponent implements OnInit, OnDestroy {
     if (context === undefined) {
       return;
     }
+    if (this.workspace && !this.workspace.entityStore.empty) {
+      this.workspace.entityStore.clear();
+    }
+    if (!this.queryState.store.empty) {
+      this.queryState.store.softClear();
+    }
+
+    this.computeHomeExtentValues(context);
 
     this.route.queryParams.pipe(debounceTime(250)).subscribe((qParams) => {
       if (!qParams['context'] || qParams['context'] === context.uri) {
@@ -1471,4 +1558,62 @@ export class PortalComponent implements OnInit, OnDestroy {
       });
     }
   }
+
+  private getFeatureIsSameActiveWks(feature: Feature): boolean {
+    if (this.workspace) {
+      const featureTitle = feature.meta.sourceTitle;
+      const wksTitle = this.workspace.title;
+      if (wksTitle === featureTitle) {
+        return true;
+      } else {
+        return false;
+      }
+    }
+    return false;
+  }
+
+  private getWksActiveOpenInResolution(): boolean {
+    if(this.workspace) {
+      const activeWks = this.workspace as WfsWorkspace;
+      if(activeWks.active && activeWks.inResolutionRange$.value && this.workspaceState.workspacePanelExpanded) {
+        return true;
+      }
+    }
+    return false;
+   }
+
+  refreshRelationsWorkspace(relationLayers: ImageLayer[] | VectorLayer[]) {
+    if (relationLayers?.length) {
+      for (const layer of relationLayers) {
+        const relationWorkspace = this.workspaceStore.all().find(workspace => layer.options.workspace.workspaceId.includes(workspace.id));
+        relationWorkspace?.meta.tableTemplate.columns.forEach(col => {
+          // Update domain list
+          if (col.type === 'list' || col.type === 'autocomplete') {
+            this.editionWorkspaceService.getDomainValues(col.relation).subscribe(result => {
+              col.domainValues = result;
+            });
+          }
+        });
+      }
+    }
+  }
+
+  zoomToSelectedFeatureWks() {
+    let format = new olFormatGeoJSON();
+    const featuresSelected = this.workspaceState.workspaceSelection.map(rec => (rec.entity as Feature));
+    if (featuresSelected.length === 0) {
+      return;
+    }
+    const olFeaturesSelected = [];
+    for (const feat of featuresSelected) {
+      let localOlFeature = format.readFeature(feat,
+        {
+          dataProjection: feat.projection,
+          featureProjection: this.map.projection
+        });
+        olFeaturesSelected.push(localOlFeature);
+    }
+    moveToOlFeatures(this.map, olFeaturesSelected, FeatureMotion.Zoom);
+  }
+
 }